+++ setup.cfg
@@ -171,7 +171,7 @@ exclude =
 
 [options.entry_points]
 console_scripts =
-    dvc = dvc.main:main
+    dvc = dvc.cli:main
 
 [flake8]
 ignore=
