+++ tests/func/machine/test_machine_status.py
@@ -1,4 +1,4 @@
-from dvc.main import main
+from dvc.cli import main
 from dvc.ui import ui
 from tests.utils import console_width
