+++ tests/func/test_move.py
@@ -3,9 +3,9 @@ import textwrap
 
 import pytest
 
+from dvc.cli import main
 from dvc.dvcfile import DVC_FILE_SUFFIX
 from dvc.exceptions import DvcException, MoveNotDataSourceError
-from dvc.main import main
 from dvc.utils.serialize import load_yaml
 from tests.basic_env import TestDvc, TestDvcGit
 from tests.func.test_repro import TestRepro
