+++ tests/func/test_root.py
@@ -1,4 +1,4 @@
-from dvc.main import main
+from dvc.cli import main
 
 
 def test_root(tmp_dir, dvc, capsys):
