+++ tests/unit/command/test_git_hook.py
@@ -16,6 +16,6 @@ def test_out_of_repo(tmp_dir, hook, cls, mocker):
     cli_args = parse_args(["git-hook", hook])
     assert cli_args.func == cls
     cmd = cli_args.func(cli_args)
-    mock_main = mocker.patch("dvc.main.main")
+    mock_main = mocker.patch("dvc.cli.main")
     assert cmd.run() == 0
     assert not mock_main.called
